backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
email-validator
sqlalchemy
aiosqlite
slack-sdk>=3.27.0
google-generativeai==0.8.5
psycopg2-binary>=2.9.9
pytz
requests
chromadb
langchain-google-genai
langchain-community
langchain
cryptography
redis>=5.0.0
fastapi-limiter>=0.1.6
tiktoken>=0.7.0
psycopg2-binary
pgvector
langchain-postgres
celery
sentry-sdk[fastapi]
alembic
pytest
pytest-asyncio
pyyaml
